ENH: Add modules for workspace configuration
Instead of hardcoding actions that are executed when entering and leaving
the workspace, the actions are defined in modules files. Each module can
define a function that is run when entering in the workspace
`${module}_activate_workspace` and a function that is run when leaving the workspace
`${module}_deactivate_workspace.

New modules can be added to extend the functionalities of this framework.