Remove `-f` argument from `type` command: bash and zsh interpret it differently
When argument `-f` is passed to Bash `type` command, it doesn't print the function
definition like Zsh does. It `suppress shell function lookup`. Instead, we can
just call the `type` function both with Zsh and Bash. With Bash, it prints the
definition of the function. With Zsh, it only returns if it is a function
or something else. Since we only care about the return code of this function
which tells us if it is a function or not, we do not care that what is
printed on screen is different, as long as the return code is the same.